# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервиса сообщений с transactional outbox.",
        "title": "Go Kafka Test API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/messages": {
            "get": {
                "description": "Возвращает сообщения с пагинацией через query-параметры offset/limit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Список сообщений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_http-server.PaginationResponse-array_github_com_xnxq1_go-kafka-test_internal_domain_Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_http-server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_http-server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт сообщение и кладёт событие в outbox в одной транзакции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Создать сообщение",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_messages.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_domain.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_http-server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_http-server.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_xnxq1_go-kafka-test_internal_domain.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "processed_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_domain.Status"
                }
            }
        },
        "github_com_xnxq1_go-kafka-test_internal_domain.Status": {
            "type": "string",
            "enum": [
                "pending",
                "sent",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusSent",
                "StatusFailed"
            ]
        },
        "github_com_xnxq1_go-kafka-test_internal_http-server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_xnxq1_go-kafka-test_internal_http-server.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "github_com_xnxq1_go-kafka-test_internal_http-server.PaginationResponse-array_github_com_xnxq1_go-kafka-test_internal_domain_Message": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_domain.Message"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_xnxq1_go-kafka-test_internal_http-server.Pagination"
                }
            }
        },
        "internal_http-server_messages.CreateMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        }
    }
}